preliminary linux support, gfxdis tool, updated gbi.h

+++ src/grc/resource.h
@@ -1,9 +1,9 @@
 #ifndef RESOURCE_H
 #define RESOURCE_H
-#include <stddef.h>
+#include <stdint.h>
 #include "grc.h"
 
 enum grc_error make_resource(const char *file_name, const char *resource_name,
-                             const void *resource_data, size_t resource_size);
+                             const void *resource_data, uint32_t resource_size);
 
 #endif
